[refactoring] moved json variable parsing in variables
